Refactor test structure for better organization
- Create test_cfi_parser.py with all CFIParser tests (23 tests)
- Reorganize test_cfi_processor.py with all CFIProcessor tests (20 tests)
- Remove confusing test_cfi_spec_compliance.py and test_cfi_range_extraction.py
- Group tests logically by the class they test rather than by feature
- Maintain same test coverage (43 tests total, all passing)

Fixes #10
